+++ magik-squid/src/main/java/nl/ramsolutions/sw/definitions/api/SwProductDefinitionGrammar.java
@@ -50,20 +50,12 @@ public enum SwProductDefinitionGrammar implements GrammarRuleKey {
             PRODUCT_IDENTIFICATION,
             builder.zeroOrMore(
                 builder.firstOf(
-                  DESCRIPTION,
-                  DO_NOT_TRANSLATE,
-                  REQUIRES,
-                  OPTIONAL,
-                  TITLE,
-                  VERSION,
-                  SPACING)),
+                    DESCRIPTION, DO_NOT_TRANSLATE, REQUIRES, OPTIONAL, TITLE, VERSION, SPACING)),
             builder.token(GenericTokenType.EOF, builder.endOfInput()));
 
     builder.rule(PRODUCT_IDENTIFICATION).is(IDENTIFIER, WHITESPACE, PRODUCT_TYPE);
     builder.rule(DESCRIPTION).is("description", SPACING, FREE_LINES, "end");
-    builder
-        .rule(DO_NOT_TRANSLATE)
-        .is("do_not_translate");
+    builder.rule(DO_NOT_TRANSLATE).is("do_not_translate");
     builder.rule(REQUIRES).is("requires", SPACING, PRODUCT_REFS, "end");
     builder.rule(OPTIONAL).is("optional", SPACING, PRODUCT_REFS, "end");
     builder.rule(TITLE).is("title", SPACING, FREE_LINES, "end");